aws/my-proj/requirements.txt
-e .
aws-cdk.aws-dynamodb
aws-cdk.aws-batch
